move fuzzer things to fuzzer module
